+++ epoch2d/src/shared_data.F90
@@ -595,7 +595,7 @@ MODULE shared_data
 
   TYPE(jobid_type) :: jobid
 
-  INTEGER :: run_date
-  INTEGER :: defines
+  INTEGER(4) :: run_date
+  INTEGER(4) :: defines
 
 END MODULE shared_data
